app/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
requests==2.28.1
python-dotenv==1.0.0
redis==4.6.0
rapidfuzz==2.10.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.10
passlib==1.7.4
python-jose==3.3.0
"pydantic[email]"
alembic
bcrypt
pytest
pytest-asyncio
mock
httpx
fastapi
starlette